Implements smarter drawing for main map view
Will only draw the invalidated rectangle passed in a "dirtyRect"

This has the potential to improve performance significantly once
features such as sprite animation and blinking are implemented. It also
makes much of the point-twiddling more readable throughout the code and
implements custom operators for CGPoint and CGSize which will remove a
lot of repetitive code.